tidy up types
