+++ .eslintrc.js
@@ -49,7 +49,7 @@ module.exports = {
 
         // variables
         "@typescript-eslint/no-unused-vars": [ "warn" ], // draw yellow line under unused vars
-        "no-undef": [ "warn" ], // draws yellow line under undefined vars
+        // "no-undef": [ "warn" ], // draws yellow line under undefined vars // it doesn't work on typescript sometimes
         "no-var": [ "error" ], // fuck you, var
         "prefer-const": [ "error" ], // const is better than let
