+++ .gitignore
@@ -1,5 +1,6 @@
 /dist
 /docs
 /types
+/ts*
 /node_modules
 **/*.log
