+++ package.json
@@ -6,6 +6,9 @@
   "repository": "https://github.com/FMS-Cat/automaton",
   "author": "FMS-Cat",
   "types": "types/index.d.ts",
+  "typesVersions": {
+    "<3.8": { "*": [ "ts3.4/*" ] }
+  },
   "license": "MIT",
   "scripts": {
     "dev": "webpack-dev-server --hot --mode development",
@@ -15,7 +18,7 @@
     "build": "yarn build-dev && yarn build-prod && yarn types",
     "build-dev": "webpack --mode development",
     "build-prod": "webpack --mode production",
-    "types": "tsc --emitDeclarationOnly",
+    "types": "tsc --emitDeclarationOnly && downlevel-dts . ts3.4",
     "docs": "typedoc --out docs --mode file --excludeNotExported",
     "lint": "eslint \"src/**/*.ts\"",
     "test": "jest"
@@ -31,6 +34,7 @@
     "@types/jest": "^25.1.4",
     "@typescript-eslint/eslint-plugin": "^2.23.0",
     "@typescript-eslint/parser": "^2.23.0",
+    "downlevel-dts": "^0.4.0",
     "eslint": "^6.8.0",
     "eslint-plugin-jest": "^23.8.2",
     "fork-ts-checker-webpack-plugin": "^4.1.0",
@@ -40,7 +44,7 @@
     "rimraf": "^3.0.2",
     "ts-jest": "^25.2.1",
     "ts-loader": "^6.2.1",
-    "typedoc": "^0.16.11",
+    "typedoc": "^0.17.6",
     "typescript": "^3.8.3",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11",
