+++ src/Channel.ts
@@ -1,36 +1,7 @@
 import { Automaton } from './Automaton';
 import { ChannelItem } from './ChannelItem';
-import { SerializedChannel } from './types/SerializedChannel';
-
-/**
- * Represent an event that is emitted by [[Channel.update]].
- */
-export interface ChannelUpdateEvent {
-  /**
-   * Current time in the current item.
-   */
-  time: number;
-
-  /**
-   * Current value of the channel.
-   */
-  value: number;
-
-  /**
-   * `true` if the update was the first call of the item.
-   */
-  init?: true;
-
-  /**
-   * `true` if the update was the last call of the item.
-   */
-  uninit?: true;
-
-  /**
-   * The progress of the item.
-   */
-  progress: number;
-}
+import type { ChannelUpdateEvent } from './types/ChannelUpdateEvent';
+import type { SerializedChannel } from './types/SerializedChannel';
 
 /**
  * It represents a channel of Automaton.
