--- src/types/index.ts
@@ -1,7 +1,0 @@
-export * from './BezierNode';
-export * from './FxDefinition';
-export * from './FxSection';
-export * from './SerializedAutomaton';
-export * from './SerializedChannel';
-export * from './SerializedChannelItem';
-export * from './SerializedCurve';
