+++ tsconfig.json
@@ -17,5 +17,5 @@
     "declarationDir": "./types"
   },
   "include": [ "src" ],
-  "exclude": [ "node_modules" ]
+  "exclude": [ "node_modules", "**/tests/**/*" ]
 }
